split DataTable files
